Moved the config server creation logic into a Task.
